Add test to native code.
Just testing out what I can do with the JNI. I've added a second
native function called mineTest that is a simple exponent of an
integer, and returns it as a double which is auto-mapped to a
jdouble. I have also found out that you can define functions in
the native source file without exposing it to the Java end.

What this means is:
1. I know how to work the JNI
2. We can essentially black box our C++. Then define a single
function that we expose to the Java side of things.