Attempt to use OpenCV4Android SDK with project
Shared libraries from the original code are in DYLIB format;
something Android can't recognise and hence code won't compile.
This was an attempt to switch to the OpenCV4Android SDK which will
fix that problem. However, the SDK doesn't contain the FaceRec
files that we needed.

This will probably be the final commit for this group project.

I really wish I was able to get this app running, as the final
chance to achieve something in uni that I'll be satisfied with.
Guess that won't be the case.

Maybe I'll clone this and work on this in the future. Who knows.